+++ files-widget/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this extension will be documented in this file.
 
+## [Unreleased]
+
+### Changed
+- Make the inline comment editor multiline with wrapped footer rendering, `Enter` for a new line, and `Ctrl+Enter` to send.
+
 ## [0.1.16] - 2026-04-19
 
 ### Fixed